created parent package for integration tests
