+++ src/it/java/it/ned/bookmanager/repository/mongo/BookMongoRepositoryIT.java
@@ -1,4 +1,4 @@
-package repository.mongo;
+package it.ned.bookmanager.repository.mongo;
 
 import it.ned.bookmanager.model.Author;
 import it.ned.bookmanager.model.Book;
@@ -9,7 +9,6 @@ import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.MongoClientSettings;
 
-import it.ned.bookmanager.repository.mongo.BookMongoRepository;
 import org.bson.codecs.configuration.CodecRegistry;
 import org.bson.codecs.pojo.PojoCodecProvider;
