+++ src/it/java/it/ned/bookmanager/service/BookServiceMongoRepositoryIT.java
@@ -1,11 +1,10 @@
-package service;
+package it.ned.bookmanager.service;
 
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
-import it.ned.bookmanager.model.Author;
-import it.ned.bookmanager.repository.mongo.AuthorMongoRepository;
-import it.ned.bookmanager.service.AuthorService;
-import it.ned.bookmanager.service.transactional.AuthorTransactionalService;
+import it.ned.bookmanager.model.Book;
+import it.ned.bookmanager.repository.mongo.BookMongoRepository;
+import it.ned.bookmanager.service.transactional.BookTransactionalService;
 import it.ned.bookmanager.transaction.TransactionManager;
 import it.ned.bookmanager.transaction.mongo.TransactionMongoManager;
 import org.junit.After;
@@ -18,20 +17,20 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 
-public class AuthorServiceMongoRepositoryIT {
+public class BookServiceMongoRepositoryIT {
 
     @ClassRule
     public static final MongoDBContainer container = new MongoDBContainer().withExposedPorts(27017);
 
-    private AuthorService service;
-    private AuthorMongoRepository repository;
+    private BookService service;
+    private BookMongoRepository repository;
     private MongoClient client;
 
     private static final String DB_NAME = "bookmanager";
     private static final String DB_BOOK_COLLECTION = "books";
     private static final String DB_AUTHOR_COLLECTION = "authors";
 
-    private static final Author AUTHOR_FIXTURE = new Author("1", "George Orwell");
+    private static final Book BOOK_FIXTURE = new Book("1", "Animal Farm", 93, "1");
 
     @Before
     public void setup() {
@@ -40,12 +39,12 @@ public class AuthorServiceMongoRepositoryIT {
 
         TransactionManager transactionManager = new TransactionMongoManager(client, DB_NAME,
                 DB_AUTHOR_COLLECTION, DB_BOOK_COLLECTION);
-        service = new AuthorTransactionalService(transactionManager);
+        service = new BookTransactionalService(transactionManager);
 
-        repository = new AuthorMongoRepository(client, client.startSession(), DB_NAME, DB_AUTHOR_COLLECTION);
+        repository = new BookMongoRepository(client, client.startSession(), DB_NAME, DB_BOOK_COLLECTION);
 
-        for (Author author: repository.findAll())
-            repository.delete(author.getId());
+        for (Book book: repository.findAll())
+            repository.delete(book.getId());
     }
 
     @After
@@ -54,21 +53,21 @@ public class AuthorServiceMongoRepositoryIT {
     }
 
     @Test
-    public void testFindAllAuthors() {
-        repository.add(AUTHOR_FIXTURE);
-        assertThat(service.findAll()).containsExactly(AUTHOR_FIXTURE);
+    public void testFindAllBooks() {
+        repository.add(BOOK_FIXTURE);
+        assertThat(service.findAll()).containsExactly(BOOK_FIXTURE);
     }
 
     @Test
-    public void testAuthorFoundWhenAdded() {
-        repository.add(AUTHOR_FIXTURE);
-        assertEquals(AUTHOR_FIXTURE, service.findById(AUTHOR_FIXTURE.getId()));
+    public void testBookFoundWhenAdded() {
+        repository.add(BOOK_FIXTURE);
+        assertEquals(BOOK_FIXTURE, service.findById(BOOK_FIXTURE.getId()));
     }
 
     @Test
-    public void testAuthorNotFoundWhenDeleted() {
-        repository.add(AUTHOR_FIXTURE);
-        repository.delete(AUTHOR_FIXTURE.getId());
-        assertNull(service.findById(AUTHOR_FIXTURE.getId()));
+    public void testBookNotFoundWhenDeleted() {
+        repository.add(BOOK_FIXTURE);
+        repository.delete(BOOK_FIXTURE.getId());
+        assertNull(service.findById(BOOK_FIXTURE.getId()));
     }
 }
